+++ azure/pre-requisite.tf
@@ -137,7 +137,7 @@ locals {
         scope = "/subscriptions/${data.azurerm_subscription.current.subscription_id}"
         role  = azurerm_role_definition.dw.name
       },
-      dw2 = {
+      dw2 = {                                                                                     // Attention: this one is not listed in document, but it is necessary
         principal_id = azurerm_user_assigned_identity.managed_id["dw"].principal_id
         scope = "/subscriptions/${data.azurerm_subscription.current.subscription_id}"
         role  = "Managed Identity Operator"